moved LogInForm and NavBar css modules to styles folder. Created logOut API call and link in the navbar - tested login/logout functionality and confirmed it's working
